refactor(cli): A.3.3 业务 inline eval-runner 到 oclif eval Command
eval-runner.ts 整段 inline 到 src/cli/oclif/commands/eval.ts（含 emitVerdict /
announceSavedReport / loadBatchChildReports 等 8 个 module-level helper）。
删 src/cli/commands/eval-runner.ts。

eval.ts 行数从 204 → ~480，oclif idiomatic Command class 内含完整业务的标准写法。
runEval 不需要 export（没有 test in-process import）。

yarn lint && build && test 1494 全绿。
