+++ src/cli/oclif/commands/eval.ts
@@ -1,13 +1,340 @@
 import { Command, Flags } from '@oclif/core';
 import { bilingual } from '../i18n.js';
 import { runLegacyCommand } from '../run-legacy.js';
+import { CliExit } from '../../cli-exit.js';
+import { tCli, type CliLang } from '../../i18n.js';
+import { parseRunConfig } from '../../parse-run-config.js';
+import { makeOnProgress } from '../../progress.js';
+import { computeRunTally } from '../../run-tally.js';
+import type { EvalArgs, EvalFlags } from '../../types/cmd-flags.js';
+import type { BatchEvaluationReport, EvaluationReport, Report, ProgressCallback } from '../../../types/index.js';
+import type { DryRunBatchReport, DryRunReport } from '../../../eval-workflows/run-evaluation.js';
+import type { EvalResult, ReportServer } from '../../_shared.js';
 
-// oclif 版 eval(默认 = run 模式) — 透传 argv 给生产 eval-runner execute()。
-// flag schema 镜像 RUN_OPTIONS(27) + eval-runner extraOptions(14) = 41 flag。
-// 描述简洁,真实语义跟约束在生产 parseRunConfig / eval-runner 里。
+// oclif 版 eval(默认 = run 模式) — 单次 typed parse 之后业务 inline。flag schema
+// 镜像 RUN_OPTIONS + eval-runner extra = 41 flag。具体语义跟约束在 parseRunConfig 里。
 //
-// `omk eval gold ...` 由 src/cli/oclif/commands/eval/gold/{init,validate,compare}.ts 处理,
-// oclif 文件目录路由自动接管,不进 eval.ts。
+// `omk eval gold ...` 由 src/cli/oclif/commands/eval/gold/{init,validate,compare}.ts
+// 处理,oclif 文件目录路由自动接管,不进 eval.ts。
+
+interface SkillProgressInfo {
+  phase: string;
+  skill: string;
+  current: number;
+  total: number;
+}
+
+interface RepeatProgressInfo {
+  run: number;
+  total: number;
+}
+
+type ParsedValues = Record<string, string | boolean | undefined>;
+
+function isDryRunReport(report: unknown): report is DryRunReport {
+  return Boolean(report && typeof report === 'object' && (report as { dryRun?: unknown }).dryRun === true);
+}
+
+function isDryRunBatchReport(report: unknown): report is DryRunBatchReport {
+  return isDryRunReport(report) && (report as { batch?: unknown }).batch === true;
+}
+
+function verdictOptions(values: ParsedValues): { gateThreshold: number; triviallySmallDiff: number | undefined } {
+  const rawThreshold = values.threshold as string | undefined;
+  const gateThreshold = rawThreshold !== undefined && Number.isFinite(Number(rawThreshold))
+    ? Number(rawThreshold)
+    : 3.5;
+  const rawTrivial = values['trivial-diff'] as string | undefined;
+  const triviallySmallDiff = rawTrivial !== undefined && Number.isFinite(Number(rawTrivial))
+    ? Number(rawTrivial)
+    : undefined;
+  return { gateThreshold, triviallySmallDiff };
+}
+
+function verdictPasses(level: string, headline: string): boolean {
+  return level === 'PROGRESS' || (level === 'SOLO' && headline.includes('PASS'));
+}
+
+function reportOnlyMode(values: ParsedValues): boolean {
+  return values['report-only'] === true || values['no-gate'] === true;
+}
+
+function applyGateExitCode(code: number, values: ParsedValues, lang: CliLang): number {
+  if (!reportOnlyMode(values)) return code;
+  process.stderr.write(tCli('cli.run.report_only_gate_skipped', lang));
+  return 0;
+}
+
+async function emitEvaluationVerdict(report: EvaluationReport, values: ParsedValues): Promise<number> {
+  const { computeVerdict, formatVerdictText } = await import('../../../eval-core/verdict.js');
+  const result = computeVerdict(report, verdictOptions(values));
+  console.log(formatVerdictText(result, { verbose: true }));
+  return verdictPasses(result.level, result.headline) ? 0 : 1;
+}
+
+function batchItemFallbackReport(
+  batch: BatchEvaluationReport,
+  item: BatchEvaluationReport['items'][number],
+): EvaluationReport {
+  return {
+    kind: 'evaluation',
+    id: item.reportId,
+    meta: {
+      ...batch.meta,
+      variants: ['baseline', item.name],
+      sampleCount: item.sampleCount,
+      totalCostUSD: item.totalCostUSD,
+      artifactHashes: item.artifactHash ? { [item.name]: item.artifactHash } : {},
+    },
+    summary: item.summary,
+    results: [],
+    ...(item.variance ? { variance: item.variance } : {}),
+  } as EvaluationReport;
+}
+
+async function loadBatchChildReports(
+  batch: BatchEvaluationReport,
+  reportsDir: string,
+  lang: CliLang,
+): Promise<EvaluationReport[]> {
+  const { createFileStore } = await import('../../../server/report-store.js');
+  const store = createFileStore(reportsDir);
+  const reports: EvaluationReport[] = [];
+  for (const item of batch.items) {
+    const loaded = await store.get(item.reportId);
+    if (loaded?.kind === 'evaluation') {
+      reports.push(loaded);
+    } else {
+      process.stderr.write(tCli('cli.run.batch_child_report_missing', lang, { id: item.reportId }));
+      reports.push(batchItemFallbackReport(batch, item));
+    }
+  }
+  return reports;
+}
+
+async function emitBatchVerdict(
+  report: BatchEvaluationReport,
+  reportsDir: string,
+  values: ParsedValues,
+  lang: CliLang,
+): Promise<number> {
+  const { computeVerdict } = await import('../../../eval-core/verdict.js');
+  const childReports = await loadBatchChildReports(report, reportsDir, lang);
+  const results = childReports.map((child) => ({
+    id: child.id,
+    treatment: child.meta.variants[1] ?? child.id,
+    verdict: computeVerdict(child, verdictOptions(values)),
+  }));
+  const passed = results.filter((r) => verdictPasses(r.verdict.level, r.verdict.headline)).length;
+  const failed = results.length - passed;
+
+  const status = lang === 'zh'
+    ? (failed === 0 ? '通过' : '未通过')
+    : (failed === 0 ? 'PASS' : 'FAIL');
+  console.log(tCli('cli.run.batch_verdict_header', lang, {
+    status,
+    passed,
+    total: results.length,
+  }));
+  for (const result of results) {
+    console.log(`  ${result.verdict.level}: ${result.treatment} — ${result.verdict.headline}`);
+  }
+  return failed === 0 ? 0 : 1;
+}
+
+async function announceSavedReport({
+  report,
+  filePath,
+  reportsDir,
+  values,
+  lang,
+}: {
+  report: EvaluationReport | BatchEvaluationReport;
+  filePath: string;
+  reportsDir: string;
+  values: ParsedValues;
+  lang: CliLang;
+}): Promise<void> {
+  const tally = computeRunTally(report);
+  process.stderr.write(tCli(report.kind === 'batch-evaluation' ? 'cli.run.batch_complete' : 'cli.run.eval_complete', lang));
+  process.stderr.write(tCli('cli.run.tally', lang, tally));
+  process.stderr.write(tCli('cli.run.report_saved', lang, { path: filePath }));
+
+  if (!values['no-serve'] && process.stdout.isTTY) {
+    const { createReportServer } = await import('../../../server/report-server.js');
+    const server: ReportServer = createReportServer({ reportsDir });
+    const serverUrl: string = await server.start();
+    const reportUrl: string = `${serverUrl}/reports/${report.id}`;
+    process.stderr.write(tCli('cli.run.report_server_running', lang, { url: serverUrl }));
+    process.stderr.write(tCli('cli.run.report_server_view', lang, { url: reportUrl }));
+    process.stderr.write(tCli('cli.run.report_server_stop', lang));
+
+    const { platform } = await import('node:os');
+    const openCmd: string = platform() === 'darwin' ? 'open' : platform() === 'win32' ? 'start' : 'xdg-open';
+    const { execFile: execFileCb } = await import('node:child_process');
+    execFileCb(openCmd, [reportUrl], () => undefined);
+  } else if (!values['no-serve']) {
+    process.stderr.write(tCli('cli.run.no_serve_in_non_tty', lang));
+    process.stderr.write(tCli('cli.run.no_serve_view_hint', lang, { id: report.id, dir: reportsDir }));
+  }
+}
+
+async function runEval(
+  _args: EvalArgs,
+  flags: EvalFlags,
+  lang: CliLang,
+): Promise<void> {
+  const { values, config, evalConfig } = parseRunConfig({ ...flags } as Record<string, unknown>);
+
+  const { runEvaluation, runMultiple, runBatchEvaluation } = await import('../../../eval-workflows/run-evaluation.js');
+
+  if (values.blind !== undefined) {
+    config.blind = values.blind as boolean | undefined;
+  }
+  config.onProgress = makeOnProgress(lang) as unknown as ProgressCallback;
+
+  const repeatRaw = values.repeat as string | undefined;
+  const parsedRepeat = repeatRaw !== undefined ? Number(repeatRaw) : (evalConfig?.repeat ?? 1);
+  if (repeatRaw !== undefined && (!Number.isFinite(parsedRepeat) || parsedRepeat < 1)) {
+    process.stderr.write(tCli('cli.run.invalid_repeat', lang, { value: repeatRaw }));
+  }
+  const repeatCount: number = Math.max(1, Math.floor(parsedRepeat) || 1);
+
+  const judgeRepeatRaw = values['judge-repeat'] as string | undefined;
+  const parsedJudgeRepeat = judgeRepeatRaw !== undefined ? Number(judgeRepeatRaw) : (evalConfig?.judgeRepeat ?? 1);
+  if (judgeRepeatRaw !== undefined && (!Number.isFinite(parsedJudgeRepeat) || parsedJudgeRepeat < 1)) {
+    process.stderr.write(tCli('cli.run.invalid_judge_repeat', lang, { value: judgeRepeatRaw }));
+  }
+  const judgeRepeatCount: number = Math.max(1, Math.floor(parsedJudgeRepeat) || 1);
+  if (judgeRepeatCount > 1) config.judgeRepeat = judgeRepeatCount;
+
+  const budgetUSD = values['budget-usd'] != null ? Number(values['budget-usd']) : undefined;
+  const budgetPerSampleUSD = values['budget-per-sample-usd'] != null ? Number(values['budget-per-sample-usd']) : undefined;
+  const budgetPerSampleMs = values['budget-per-sample-ms'] != null ? Number(values['budget-per-sample-ms']) : undefined;
+  if (budgetUSD !== undefined || budgetPerSampleUSD !== undefined || budgetPerSampleMs !== undefined) {
+    config.budget = {
+      ...(budgetUSD !== undefined && Number.isFinite(budgetUSD) && budgetUSD >= 0 ? { totalUSD: budgetUSD } : {}),
+      ...(budgetPerSampleUSD !== undefined && Number.isFinite(budgetPerSampleUSD) && budgetPerSampleUSD >= 0 ? { perSampleUSD: budgetPerSampleUSD } : {}),
+      ...(budgetPerSampleMs !== undefined && Number.isFinite(budgetPerSampleMs) && budgetPerSampleMs >= 0 ? { perSampleMs: budgetPerSampleMs } : {}),
+    };
+  }
+
+  const lengthDebiasOff = (values['no-debias-length'] as boolean | undefined) === true
+    || (values['no-debias-length'] === undefined && evalConfig?.lengthDebias === false);
+  if (lengthDebiasOff) {
+    config.lengthDebias = false;
+    process.stderr.write(tCli('cli.run.no_debias_length_active', lang));
+  }
+
+  const bootstrapEnabled = (values.bootstrap as boolean | undefined) === true
+    || (values.bootstrap === undefined && evalConfig?.bootstrap !== false);
+  if (bootstrapEnabled) {
+    config.bootstrap = true;
+    const bsRaw = values['bootstrap-samples'] as string | undefined;
+    const parsedBs = bsRaw !== undefined ? Number(bsRaw) : (evalConfig?.bootstrapSamples ?? 1000);
+    if (bsRaw !== undefined && (!Number.isFinite(parsedBs) || parsedBs < 100)) {
+      process.stderr.write(tCli('cli.run.invalid_bootstrap_samples', lang, { value: bsRaw }));
+    }
+    const bsCount = Math.max(100, Math.floor(parsedBs) || 1000);
+    if (bsCount > 10000) {
+      process.stderr.write(tCli('cli.run.bootstrap_samples_too_large', lang, { n: bsCount }));
+    }
+    config.bootstrapSamples = bsCount;
+  }
+
+  config.lang = lang;
+  if (values['skip-connectivity'] as boolean) {
+    process.stderr.write(tCli('cli.run.skip_connectivity_warning', lang) + '\n');
+  }
+
+  try {
+    if (values.batch) {
+      const { report, filePath } = await runBatchEvaluation({
+        ...config,
+        repeat: repeatCount,
+        onSkillProgress({ phase, skill, current, total }: SkillProgressInfo): void {
+          if (phase === 'start') {
+            process.stderr.write(tCli('cli.run.skill_section', lang, {
+              i: current ?? '', n: total ?? '', skill: skill ?? '',
+            }));
+          }
+        },
+      }) as { report: BatchEvaluationReport | DryRunBatchReport; filePath: string | null };
+      console.log(JSON.stringify(report, null, 2));
+      if (isDryRunBatchReport(report)) {
+        console.log(tCli('cli.run.dry_run_no_scores', lang));
+        throw new CliExit(0);
+      }
+      if (filePath) {
+        await announceSavedReport({ report, filePath, reportsDir: config.outputDir, values, lang });
+      }
+      const exitCode = await emitBatchVerdict(report, config.outputDir, values, lang);
+      throw new CliExit(applyGateExitCode(exitCode, values, lang));
+    }
+
+    let report: Report;
+    let filePath: string | null;
+
+    if (repeatCount > 1 && !config.dryRun) {
+      const result = await runMultiple({
+        ...config,
+        repeat: repeatCount,
+        onRepeatProgress({ run, total }: RepeatProgressInfo): void {
+          process.stderr.write(tCli('cli.run.run_section', lang, { i: run, n: total }));
+        },
+      }) as { report: Report };
+      report = result.report;
+      filePath = null;
+    } else {
+      const result = (await runEvaluation({
+        ...config,
+        ...(repeatCount > 1 ? { repeat: repeatCount } : {}),
+      })) as EvalResult;
+      if (isDryRunReport(result.report)) {
+        console.log(JSON.stringify(result.report, null, 2));
+        console.log(tCli('cli.run.dry_run_no_scores', lang));
+        throw new CliExit(0);
+      }
+      report = result.report as Report;
+      filePath = result.filePath;
+    }
+
+    const goldDir = (values['gold-dir'] as string | undefined) ?? evalConfig?.goldDir;
+    if (goldDir && filePath) {
+      const { attachGoldAgreementToReport, formatGoldCompare } = await import('../../../grading/gold-cli.js');
+      const out = attachGoldAgreementToReport({
+        report,
+        goldDir,
+        outputDir: config.outputDir,
+        samples: config.bootstrapSamples,
+      });
+      if (out.result && out.gold) {
+        process.stderr.write(formatGoldCompare(out.result, out.gold));
+        if (out.result.contaminationWarning) {
+          process.stderr.write(tCli('cli.run.contamination_warning', lang, {
+            warning: out.result.contaminationWarning,
+          }));
+        }
+      } else {
+        process.stderr.write(tCli('cli.run.gold_load_failed', lang, { dir: goldDir }));
+        for (const m of out.loadIssues) {
+          process.stderr.write(tCli('cli.run.gold_load_issue', lang, { message: m }));
+        }
+      }
+    }
+
+    console.log(JSON.stringify(report, null, 2));
+    if (filePath) {
+      await announceSavedReport({ report, filePath, reportsDir: config.outputDir, values, lang });
+    }
+    const exitCode = await emitEvaluationVerdict(report, values);
+    throw new CliExit(applyGateExitCode(exitCode, values, lang));
+  } catch (err: unknown) {
+    if (err instanceof CliExit) throw err;
+    console.error(tCli('cli.common.error_prefix', lang, { message: (err as Error).message }));
+    throw new CliExit(1);
+  }
+}
 
 export default class Eval extends Command {
   static description = bilingual({
@@ -197,7 +524,6 @@ export default class Eval extends Command {
     const { args, flags } = await this.parse(Eval);
     const lang = (flags.lang ?? 'zh') as 'zh' | 'en';
     await runLegacyCommand(this, async () => {
-      const { runEval } = await import('../../commands/eval-runner.js');
       await runEval(args as Record<string, never>, { ...flags, lang }, lang);
     });
   }
